% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd2q-package.R
\docType{package}
\name{bd2q-package}
\alias{bd2q}
\alias{bd2q-package}
\title{bd2q: Convert a 'blogdown' Blog to Quarto}
\description{
Takes the folder stucture and files of an old 'blogdown' blog and translates them to what's needed for a Quarto blog. Built for my own personal blog, so your mileage may vary.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/bd2q}
  \item Report bugs at \url{https://github.com/matt-dray/bd2q/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
