% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer-blogdown-to-quarto.R
\name{create_and_transfer}
\alias{create_and_transfer}
\title{Transfer a 'blogdown' Blog to Quarto Structure}
\usage{
create_and_transfer(
  bd_path,
  q_path,
  rproj = TRUE,
  resources_dir = "resources",
  exts_keep = c("gif", "jpg", "jpeg", "png", "svg", "wav"),
  overwrite = FALSE
)
}
\arguments{
\item{bd_path}{Character. Path to directory containing a 'blogdown' blog from
where you want to copy posts}

\item{q_path}{Character. Path to directory containing a Quarto blog where you
want to copy posts to.}

\item{rproj}{Logical. Include an R Project (.Rproj) file? Defaults to
\code{TRUE}.}

\item{resources_dir}{Character. The name of the folder that contains
resources, such as images, that are used in posts.}

\item{exts_keep}{Character. A vector of extensions for files to keep from the
'blogdown' blog.}

\item{overwrite}{Logical. Overwrite the existing directory and all posts and
resource files at the provided q_path? Defaults to \code{FALSE}.}
}
\value{
Nothing. New directory structure and files are created at the path
given by q_path.
}
\description{
A convenience function that creates a Quarto blog template and transfers the
structure and content of posts and resources from a 'blogdown' blog project.
}
\details{
Executes \link{create_template}, \link{transfer_posts} and \link{transfer_resources} to
set up the basic Quarto blog template and generate and fill the posts/
directory based on the posts found in the provided 'blogdown' blog project
}
\examples{
\dontrun{create_and_transfer}
}
