% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer-blogdown-to-quarto.R
\name{create_template}
\alias{create_template}
\title{Generate a Quarto Blog Template}
\usage{
create_template(path, rproj = TRUE, overwrite = FALSE)
}
\arguments{
\item{path}{Character. Path to a directory where you want to create a Quarto
blog.}

\item{rproj}{Logical. Include an R Project (.Rproj) file? Defaults to
\code{TRUE}.}

\item{overwrite}{Logical. Overwrite an existing directory at the provided
path? Defaults to \code{FALSE}.}
}
\value{
Nothing. New directory structure and files are created at the path
provided by the user.
}
\description{
Create template structure and files for a basic Quarto blog.
}
\details{
Creates a blog directory at the user-specified path, with the files
_quarto.yml, about.qmd, index.qmd, posts/_metadata.yml, styles.css and
an optional .Rproj file. This is similar to the starter blog created in
\href{https://quarto.org/docs/websites/website-blog.html}{the Quarto quickstart guide},
but with some opinionated differences.
}
\examples{
\dontrun{create_template("~/Documents/new-quarto-project")}
}
