% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{delete_draft_line}
\alias{delete_draft_line}
\title{Remove 'Draft' Status from Quarto YAML}
\usage{
delete_draft_line(q_path)
}
\arguments{
\item{q_path}{Character. Path to directory containing the Quarto blog.}
}
\value{
Nothing. Quarto files are overwritten if they contained a draft
status in their YAML header.
}
\description{
Search for the YAML line in each post's index.qmd that might contain the
string 'draft:', remove them and overwrite the original file.
}
\details{
A qmd with 'demo: no' in the YAML will render, but will be skipped
when it comes to the site's index.qmd being rendered; these posts simply
won't appear on the homepage. We can read in the qmd, remove the line
with 'draft:' if it exists, and resave it.
}
\examples{
\dontrun{function("~/Documents/matt-dray/rostrum-blog-2")}
}
