% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-qmd-files.R
\name{remove_line}
\alias{remove_line}
\title{Remove a Line from a Quarto Post}
\usage{
remove_line(q_path, detect_rx)
}
\arguments{
\item{q_path}{Character. Path to directory containing the Quarto blog.}

\item{detect_rx}{Character. A regular expression that captures text within a
line that you want to delete from each post's index.qmd file.}
}
\value{
Nothing. Quarto files are overwritten if they contain matching lines.
}
\description{
Search each Quarto blog post's index.qmd file for lines that contain a
supplied regular expression, delete them and overwrite the original file.
}
\examples{
\dontrun{
remove_line(
  q_path = "~/Documents/new-quarto-project/",
  detect_rx = "^draft:"  # remove YAML line that specifies draft status
)
}
}
