% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-qmd-files.R
\name{remove_lines}
\alias{remove_lines}
\title{Remove Lines from a Quarto Post}
\usage{
remove_lines(q_path, detect_strs)
}
\arguments{
\item{q_path}{Character. Path to directory containing the Quarto blog.}

\item{detect_strs}{Character. A vector of strings that matches text within a
line that you want to delete from each post's index.qmd file.}
}
\value{
Nothing. Quarto files are overwritten if they contain matching lines.
}
\description{
Search each Quarto blog post's index.qmd file for consecutive lines that
contain a supplied set of strings, delete them and overwrite the original
file.
}
\examples{
\dontrun{
remove_lines(
  q_path = "~/Documents/new-quarto-project/",
  detect_strs = c(  # remove bespoke session info block
    "---",
    "```{r}",
    "sessionInfo()",
    "```"
  )
)
}
}
