% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-qmd-files.R
\name{replace_lines}
\alias{replace_lines}
\title{Replace or Remove Lines from a Post}
\usage{
replace_lines(q_path, match_str, replacement_str, collapse_str = "///")
}
\arguments{
\item{q_path}{Character. Path to directory containing the Quarto blog.}

\item{match_str}{Character. A vector of strings that match exactly to
consecutive lines in a blog post that you want to replace (with
'replacement_str') or remove (when 'replacement_str' is \code{NULL}).}

\item{replacement_str}{Character. A vector of consecutive strings to replace
the strings proviede in 'match_str'. Specify \code{NULL} (default) if you want
to remove 'match_str' without replacement.}

\item{collapse_str}{Character. The function works by collapsing a post's
lines to a single string, with elements separated by some unique string.}
}
\value{
Nothing. Quarto files are overwritten if they contain matching lines.
}
\description{
Replace or Remove Lines from a Post
}
\examples{
\dontrun{
replace_lines(
  q_path = "~/Documents/new-quarto-project/",
  match_str = c("This is an example", "These are consecutive lines to match"),
  replacement_str = c("Replace with this line", "And this one")
)
}
}
