% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer-blogdown-to-quarto.R
\name{transfer_posts}
\alias{transfer_posts}
\title{Transfer Posts from 'blogdown' to Quarto}
\usage{
transfer_posts(bd_path, q_path, ...)
}
\arguments{
\item{bd_path}{Character. Path to directory containing a 'blogdown' blog from
where you want to copy posts}

\item{q_path}{Character. Path to directory containing a Quarto blog where you
want to copy posts to.}

\item{...}{Passed to \link[fs:copy]{fs::file_copy}, with the intention that you can supply
\code{overwrite = TRUE} if you need to overwrite any existing files.}
}
\value{
Nothing. New directory structure and files are created at the path
given by q_path.
}
\description{
Generate the folder structure required to move R Markdown (.Rmd) posts from a
'blogdown' project to a Quarto project and then copy them over, naming each
one 'index.qmd' in its own folder.
}
\details{
Assumes that Rmd posts in the 'blogdown' blog are at the path
/content/post/YYYY-MM-DD-post-name.Rmd. For Quarto, the corresponding path
will be /posts/YYYY-MM-DD-post-name/ and the file containing each post
will be named index.qmd. (Each of these individual post directories can
also hold a folder of  associated resources for each post, which can be
generated and copied from a 'blogdown' blog project using the
\link{transfer_resources} function.)
}
\examples{
\dontrun{
transfer_posts(
    bd_path = "~/Documents/old-blogdown-project/",
    q_path  = "~/Documents/new-quarto-project/",
    overwrite = TRUE
)
}
}
