% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{transfer_resources}
\alias{transfer_resources}
\title{Add Resources to Quarto Post Directories}
\usage{
transfer_resources(
  bd_path,
  q_path,
  resources_dir = "resources",
  exts_keep = c("gif", "jpg", "jpeg", "png", "svg", "wav"),
  ...
)
}
\arguments{
\item{bd_path}{Character. Path to directory containing the 'blogdown' blog.}

\item{q_path}{Character. Path to directory containing the Quarto blog.}

\item{resources_dir}{Character. The name of the folder that contains
resources, such as images, that are used in posts.}

\item{exts_keep}{Character. A vector of extensions for files to keep from the
blogdown blog.}

\item{...}{Passed to \link[fs:copy]{fs::file_copy}, with the intention that you can supply
\code{overwrite = TRUE} if you need to overwrite any existing files.}
}
\value{
Nothing. New directory structure and files are created at the path
given by q_path, in the subdirectory given by resources_dir.
}
\description{
Copy resources (images, etc) from the directory structure of a 'blogdown'
blog to that required by Quarto. Assumes you've already set up the directory
structure for each post using \link{transfer_posts}.
}
\details{
Resources for each post in the 'blogdown' blog are at the path
/static/post/YYYY-MM-DD-post-name_files/. For Quarto, the folder of
resources for each post lives under the appropriate
/posts/YYYY-MM-DD-post-name/ path alongside it's associated index.qmd.
}
\examples{
\dontrun{
transfer_resources(
    bd_path = "~/Documents/matt-dray/rostrum-blog",
    q_path  = "~/Documents/matt-dray/rostrum-blog-2",
    resources_dir = "resources",
    exts_keep = c("gif", "jpg", "jpeg", "png", "svg", "wav"),
    overwrite = TRUE
)
}
}
