% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-qmd-files.R
\name{replace_lines}
\alias{replace_lines}
\title{Replace or Remove Lines from a Post}
\usage{
replace_lines(
  q_path = "/Users/mattdray/Desktop/test",
  match_str = c("---", "<details><summary>Session info</summary>",
    "```{r eval=TRUE, sessioninfo, echo=FALSE}", "sessioninfo::session_info()", "```",
    "</details>", ""),
  replacement_str = "hello!",
  collapse_str = "///"
)
}
\arguments{
\item{q_path}{Character. Path to directory containing the Quarto blog.}

\item{match_str}{Character. A vector of consecutive lines in a post that you
want to match, so they can be replaced or removed. This is based on exact
matching, not a regular expression.}

\item{replacement_str}{Character. The consecutive lines that you want to
insert to replace the vector provided by 'match_str'. Specify \code{NULL}
(default) if you want to remove match_str' without replacement.}

\item{collapse_str}{Character. The function works by collapsing a post's
lines to a single string, with elements separated by some unique string.}
}
\value{
Nothing. Quarto files are overwritten if they contain matching lines.
}
\description{
Replace or Remove Lines from a Post
}
\examples{
\dontrun{
replace_lines(
  q_path = "~/Documents/new-quarto-project/",
  match_str = c("This is an example", "These are consecutive lines to match"),
  replacement_str = c("Replace with this line", "And this one"),
  collapse_str = "||||"
)
}
}
