% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{transfer_posts}
\alias{transfer_posts}
\title{Build Directory Structure and Add Posts}
\usage{
transfer_posts(bd_path, q_path)
}
\arguments{
\item{bd_path}{Character. Path to directory containing the 'blogdown' blog.}

\item{q_path}{Character. Path to directory containing the Quarto blog.}
}
\value{
Nothing. New directory structure and files are created at the path
given by q_path.
}
\description{
Convert the directory structure for existing 'blogdown' posts to that needed
for a newly-created Quarto blog.
}
\details{
Rmd posts in the 'blogdown' blog are at the path
/content/post/YYYY-MM-DD-post-name.Rmd. For Quarto, the path must be
/posts/YYYY-MM-DD-post-name/ and the file containing each post must be
named 'index.qmd'. This directory will also hold a folder of associated
resources, which can be generated using \link{transfer_resources}.
}
\examples{
\dontrun{
transfer_posts(
    bd_path = "~/Documents/matt-dray/rostrum-blog",
    q_path  = "~/Documents/matt-dray/rostrum-blog-2"
)
}
}
