% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer-blogdown-to-quarto.R
\name{transfer_resources}
\alias{transfer_resources}
\title{Transfer Resources from 'blogdown' to Quarto}
\usage{
transfer_resources(
  bd_path,
  q_path,
  resources_dir = "resources",
  exts_keep = c("gif", "jpg", "jpeg", "png", "svg", "wav"),
  ...
)
}
\arguments{
\item{bd_path}{Character. Path to directory containing a 'blogdown' blog from
where you want to copy resources.}

\item{q_path}{Character. Path to directory containing a Quarto blog where you
want to copy resources to.}

\item{resources_dir}{Character. The name of the folder that contains
resources, such as images, that are used in posts.}

\item{exts_keep}{Character. A vector of extensions for files to keep from the
'blogdown' blog.}

\item{...}{Passed to \link[fs:copy]{fs::file_copy}, with the intention that you can supply
\code{overwrite = TRUE} if you need to overwrite any existing files.}
}
\value{
Nothing. New directory structure and files are created at the path
given by q_path, in the subdirectory given by resources_dir.
}
\description{
Copy resources (images, audio files, etc) from the directory structure of a
'blogdown' blog to that required by Quarto. Assumes you've already set up the
directory structure for each post using \link{transfer_posts}.
}
\details{
Assumes that the resources for a given 'blogdown' post are at the
path /static/post/YYYY-MM-DD-post-name_files/, and the that the
corresponding Quarto project will have a subdirectory for resources, which
defaults to the name 'resources/', in the folder
/posts/YYYY-MM-DD-post-name/ alongside that post's index.qmd file.
}
\examples{
\dontrun{
transfer_resources(
    bd_path = "~/Documents/old-blogdown-project/",
    q_path  = "~/Documents/new-quarto-project/",
    resources_dir = "resources",
    exts_keep = c("gif", "jpg", "jpeg", "png", "svg", "wav"),
    overwrite = TRUE
)
}
}
